刚才还是有没改到的地方

+++ server/include/utility2.h
@@ -6,7 +6,7 @@
 class utility2
 {
 public:
-	static std::string toIPs(unsigned long IPu)
+	static std::string toIPs(uint32_t IPu)
 	{
 		return boost::asio::ip::address_v4(IPu).to_string();
 	}
